+++ src/test/java/io/tarantool/driver/integration/ConvertersWithClusterClientIT.java
@@ -15,6 +15,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIf;
 
+import java.time.Instant;
 import java.util.UUID;
 import org.testcontainers.shaded.org.apache.commons.lang.ArrayUtils;
 
@@ -63,6 +64,23 @@ public class ConvertersWithClusterClientIT extends SharedTarantoolContainer {
 
         //then
         Assertions.assertEquals(uuid, fields.getUUID("uuid_field"));
+    }
+
+    @Test
+    @EnabledIf("io.tarantool.driver.TarantoolUtils#versionWithInstant")
+    public void test_boxSelect_shouldReturnTupleWithInstant() throws Exception {
+        //given
+        Instant instant = Instant.now();
+        client.space("space_with_instant")
+                .insert(tupleFactory.create(1, instant)).get();
+
+        //when
+        TarantoolTuple fields = client
+                .space("space_with_instant")
+                .select(Conditions.equals("id", 1)).get().get(0);
+
+        //then
+        Assertions.assertEquals(instant, fields.getInstant("instant_field"));
     }
 
     @Test