+++ src/test/java/io/tarantool/driver/integration/ConvertersWithClusterClientIT.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIf;
 
 import java.util.UUID;
-import org.testcontainers.shaded.org.apache.commons.lang.ArrayUtils;
+import org.testcontainers.shaded.org.apache.commons.lang3.ArrayUtils;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;