+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Adding default mapper for long arrays ([#290](https://github.com/tarantool/cartridge-java/pull/290))
+- Support Datetime type ([#293](https://github.com/tarantool/cartridge-java/pull/293))
 
 ## [0.9.1] - 2022-10-13