+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adding default mapper for long arrays ([#290](https://github.com/tarantool/cartridge-java/pull/290))
 - Support Datetime type ([#293](https://github.com/tarantool/cartridge-java/pull/293))
 
+
 ## [0.9.1] - 2022-10-13
 
 - Changed TarantoolNullField class to singleton ([#195](https://github.com/tarantool/cartridge-java/pull/275))